+++ lib/src/file_storage.dart
@@ -1,15 +1,22 @@
 import 'dart:io';
 import 'dart:typed_data';
+import 'package:path/path.dart';
 import 'package:path_provider/path_provider.dart';
 
 class FileStorage {
-  static Future<String> get _localPath async =>
-      (await getApplicationDocumentsDirectory()).path;
+  String? outputPath;
 
-  static Future<File> _localFile(String filename) async =>
-      File('${await _localPath}/$filename');
+  FileStorage();
 
-  static Future<Uint8List?> read(String filename) async {
+  Future<String> get documentsPath async => (await getApplicationDocumentsDirectory()).path;
+
+  Future<File> _localFile(String filename) async {
+    final outputPath = this.outputPath;
+    if (outputPath == null) return File(join(await documentsPath, filename));
+    return File(join(await documentsPath, outputPath, filename));
+  }
+
+  Future<Uint8List?> read(String filename) async {
     try {
       final file = await _localFile(filename);
       return file.readAsBytes();
@@ -18,23 +25,26 @@ class FileStorage {
     }
   }
 
-  static Future<bool> exists(String filename) async {
+  Future<bool> exists(String filename) async {
     try {
       final file = await _localFile(filename);
-      return file.exists();
+      return file.existsSync();
     } catch (e) {
       return false;
     }
   }
 
-  static Future<File> write(String filename, Uint8List content) async {
+  Future<File> write(String filename, Uint8List content) async {
     final file = await _localFile(filename);
+    if (!file.existsSync()) {
+      await file.create(recursive: true);
+    }
     return file.writeAsBytes(content);
   }
 
-  static Future<void> delete(String filename) async {
+  Future<void> delete(String filename) async {
     final file = await _localFile(filename);
-    if (!await file.exists()) return;
+    if (!file.existsSync()) return;
     await file.delete();
   }
 }