+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.3 - 06/01/2022
+
+### Fixed
+* bug where the `_keys` were not populated correctly (keys that were used before 0.0.3 will not be populated)
+### Added
+* Support for setting a custom output path
+
 ## 0.0.2 - 17/12/2021
 
 * encrypt and decrypt native on Android and iOS > 12